requirements.txt
pillow

# Django Packages
django-imagekit
django-cleanup